Addressing signed vs unsigned comparison warning in assert
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1825 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Misc/ActionQueues.cpp
@@ -100,7 +100,7 @@ ActionQueues::reset()
 void
 ActionQueues::resetQueue(int inPlayerIndex)
 {
-	assert(inPlayerIndex >= 0 && inPlayerIndex < mNumPlayers);
+	assert(inPlayerIndex >= 0 && inPlayerIndex < static_cast<int>(mNumPlayers));
 	mQueueHeaders[inPlayerIndex].read_index = mQueueHeaders[inPlayerIndex].write_index = 0;
 }
